Separate CLI, framework and targets
