+++ kevm-pyk/pyproject.toml
@@ -42,7 +42,7 @@ pyupgrade = "*"
 [tool.poetry.scripts]
 kevm = "kevm_pyk.__main__:main"
 kevm-pyk = "kevm_pyk.__main__:main"
-kevm-dist = "kevm_pyk.dist:main"
+kevm-dist = "kevm_pyk.kdist.__main__:main"
 gst-to-kore = "kevm_pyk.gst_to_kore:main"
 kontrolx = "kontrolx.__main__:main"
