+++ kevm-pyk/src/kevm_pyk/interpreter.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 from pyk.kore.parser import KoreParser
 from pyk.utils import run_process
 
-from . import dist
+from . import kdist
 from .gst_to_kore import gst_to_kore
 
 if TYPE_CHECKING:
@@ -26,7 +26,7 @@ def interpret(gst_data: Any, schedule: str, mode: str, chainid: int, *, check: b
 
 
 def _interpret(gst_data: Any, schedule: str, mode: str, chainid: int) -> CompletedProcess:
-    interpreter = dist.get('llvm') / 'interpreter'
+    interpreter = kdist.get('llvm') / 'interpreter'
     init_kore = gst_to_kore(gst_data, schedule, mode, chainid)
     proc_res = run_process([str(interpreter), '/dev/stdin', '-1', '/dev/stdout'], input=init_kore.text, check=False)
     return proc_res
