+++ kevm-pyk/src/tests/integration/test_run.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 import pytest
 from pyk.kore.tools import PrintOutput, kore_print
 
-from kevm_pyk import dist
+from kevm_pyk import kdist
 from kevm_pyk.interpreter import interpret
 
 from ..utils import REPO_ROOT
@@ -32,7 +32,7 @@ def test_run(gst_file: Path) -> None:
 
     # When
     pattern = interpret(gst_data, 'SHANGHAI', 'NORMAL', 1, check=False)
-    actual = kore_print(pattern, definition_dir=dist.get('llvm'), output=PrintOutput.PRETTY)
+    actual = kore_print(pattern, definition_dir=kdist.get('llvm'), output=PrintOutput.PRETTY)
 
     # Then
     assert actual == expected
