+++ kevm-pyk/src/tests/integration/utils.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 
 from pyk.utils import check_dir_path
 
-from kevm_pyk import dist
+from kevm_pyk import kdist
 from kontrolx.solc_to_k import solc_to_k
 
 if TYPE_CHECKING:
@@ -23,7 +23,7 @@ def gen_bin_runtime(contract_file: Path, output_dir: Path) -> tuple[Path, str]:
     main_file = output_dir / f'{contract_name.lower()}-bin-runtime.k'
 
     k_text = solc_to_k(
-        definition_dir=dist.get('haskell'),
+        definition_dir=kdist.get('haskell'),
         contract_file=contract_file,
         contract_name=contract_name,
         main_module=main_module_name,
